requirements.txt
streamlit>=1.28.0
groq>=0.4.0
stability-sdk>=0.8.0
Pillow>=10.0.0
requests>=2.31.0
